+++ variables.tf
@@ -2,16 +2,25 @@ variable "public_ip_config" {
   type = object({
     allocation_method = optional(string, "Static")
     stage             = string
+    sku               = optional(string, "Standard")
   })
   default = null
   validation {
     condition     = var.public_ip_config != null ? contains(["Static", "Dynamic"], var.public_ip_config.allocation_method) : true
     error_message = "Allocation method must be Static or Dynamic"
   }
+
+  validation {
+    condition = (
+      var.virtual_machine_config.zone == null || var.public_ip_config.sku == "Standard"
+    )
+    error_message = "If a zone is specified, the Public IP SKU must be set to 'Standard'."
+  }
   description = <<-DOC
   ```
     allocation_method: The allocation method of the public ip that will be created. Defaults to static.
     stage: The stage of this PIP. Ex: prd, dev, tst, ...
+    sku: Optionally specify the sku of the public ip. Defaults to Standard.
   ```
   DOC
 }
@@ -30,7 +39,7 @@ variable "nic_config" {
   nullable    = false
   description = <<-DOC
   ```
-    private_ip: Optioanlly specify a private ip to use. Otherwise it will  be allocated dynamically.
+    private_ip: Optionally specify a private ip to use. Otherwise it will be allocated dynamically.
     dns_servers: Optionally specify a list of dns servers for the nic.
     enable_accelerated_networking: Enabled Accelerated networking (SR-IOV) on the NIC. The machine SKU must support this feature.
     nsg: Although it is discouraged you can optionally assign an NSG to the NIC. Optionally specify a NSG object.
@@ -48,10 +57,19 @@ variable "additional_network_interface_ids" {
 variable "subnet" {
   type = object({
     id               = string
-    address_prefixes = list(string)
+    address_prefixes = optional(list(string), null)
   })
   nullable    = false
-  description = "The variable takes the subnet as input and takes the id and the address prefix for further configuration."
+  description = <<-DOC
+  ```
+    The variable takes the subnet as input and takes the id and the address prefix for further configuration.
+    Note: If no address prefix is provided, the information is being extracted from the id.
+  ```
+  DOC
+  validation {
+    condition = var.subnet.address_prefixes == null ? can(regex(".*subnets/snet-[0-9-]+-.*$", var.subnet.id)) : true
+    error_message = "If no address prefix is specified, the name of the subnet must match the naming convention."
+  }
 }
 
 variable "admin_username" {
@@ -142,6 +160,11 @@ variable "virtual_machine_config" {
     condition     = var.virtual_machine_config.zone == null || var.virtual_machine_config.zone == "1" || var.virtual_machine_config.zone == "2" || var.virtual_machine_config.zone == "3"
     error_message = "Zone, can only be empty, 1, 2 or 3."
   }
+  validation {
+    condition     = var.virtual_machine_config.zone != null ? var.virtual_machine_config.availability_set_id == null : true
+    error_message = "Either 'zone' or 'availability_set_id' can be set, but not both."
+  }
+
   nullable    = false
   description = <<-DOC
   ```
@@ -177,8 +200,11 @@ variable "data_disks" {
     storage_account_type       = optional(string, "Premium_LRS")
     write_accelerator_enabled  = optional(bool, false)
     on_demand_bursting_enabled = optional(bool, false)
-    disk_iops_read_write       = optional(string, null)
-    disk_mbps_read_write       = optional(string, null)
+    disk_iops_read_write       = optional(number)
+    disk_mbps_read_write       = optional(number)
+    disk_iops_read_only        = optional(number)
+    disk_mbps_read_only        = optional(number)
+    max_shares                 = optional(number)
   }))
   validation {
     condition     = length([for v in var.data_disks : v.lun]) == length(distinct([for v in var.data_disks : v.lun]))
@@ -189,18 +215,17 @@ variable "data_disks" {
     error_message = "Possible values are Standard_LRS, StandardSSD_LRS, Premium_LRS, StandardSSD_ZRS, Premium_ZRS, PremiumV2_LRS and UltraSSD_LRS for storage_account_type"
   }
   validation {
-    condition = alltrue([for o in var.data_disks : (contains(["Standard_LRS", "StandardSSD_LRS", "Premium_LRS", "StandardSSD_ZRS", "Premium_ZRS"], o.storage_account_type) && (o.disk_iops_read_write != null || o.disk_mbps_read_write != null)
-    || (contains(["PremiumV2_LRS", "UltraSSD_LRS"], o.storage_account_type)))])
-    error_message = "disk_iops_read_write and disk_mbps_read_write are not supported for Standard_LRS, StandardSSD_LRS, Premium_LRS, StandardSSD_ZRS and Premium_ZRS storage_account_type"
+    condition = alltrue([
+      for disk in var.data_disks : (
+        disk.on_demand_bursting_enabled == false ||
+        disk.disk_size_gb > 512
+      )
+    ])
+    error_message = "on_demand_bursting_enabled` can only be set to true when `disk_size_gb` is larger than 512GB."
   }
   validation {
-    condition = alltrue([for o in var.data_disks : ((contains(["PremiumV2_LRS", "UltraSSD_LRS"], o.storage_account_type) && (var.virtual_machine_config.zone != null))
-    || (contains(["Standard_LRS", "StandardSSD_LRS", "Premium_LRS", "StandardSSD_ZRS", "Premium_ZRS"], o.storage_account_type)))])
-    error_message = "Premium SSD v2 & Ultra SSD disks can only be attached to zonal VMs."
-  }
-  validation {
-    condition = alltrue([for o in var.data_disks : (
-      (o.write_accelerator_enabled == true && contains(["Premium_LRS", "Premium_ZRS"], o.storage_account_type) && contains(["None", "ReadOnly"], o.caching)) ||
+    condition = alltrue([for o in var.data_disks: (
+      (o.write_accelerator_enabled == true && contains(["Premium_LRS", "Premium_ZRS"], o.storage_account_type) && contains(["None"], o.caching)) ||
       (o.write_accelerator_enabled == false)
     )])
     error_message = "write_accelerator_enabled, can only be activated on Premium disks and caching deactivated."
@@ -210,24 +235,117 @@ variable "data_disks" {
       (o.on_demand_bursting_enabled == true && contains(["Premium_LRS", "Premium_ZRS"], o.storage_account_type)) ||
       (o.on_demand_bursting_enabled == false)
     )])
-    error_message = "If enable on demand bursting, possible storage_account_type values are Premium_LRS and Premium_ZRS."
+    error_message = "If enable on demand bursting, possible storage_account_type values are Premium_LRS, Premium_ZRS"
+  }
+  validation {
+    condition = alltrue([
+      for v in var.data_disks :
+      (
+        (v.storage_account_type != "PremiumV2_LRS") || 
+        (var.virtual_machine_config.zone != null)
+      )
+    ])
+    error_message = "PremiumV2_LRS storage_account_type requires zone to be set. Please set zone to 1, 2 or 3."
+  }
+  validation {
+    condition = (
+      length(distinct([
+        for v in var.data_disks : var.virtual_machine_config.zone
+      ])) <= 1
+    )
+    error_message = "All disks must be in the same zone or zone must be empty."
+  }
+  validation {
+    condition = alltrue([
+      for v in var.data_disks :
+      (
+        !(v.storage_account_type == "PremiumV2_LRS" || v.storage_account_type == "UltraSSD_LRS") ||
+        (v.caching == "None")
+      )
+    ])
+    error_message = "When storage_account_type is 'PremiumV2_LRS' or 'UltraSSD_LRS', caching must be set to 'None'."
   }
   validation {
     condition     = alltrue([for k, v in var.data_disks : !strcontains(k, "-")])
     error_message = "Logical Name can't contain a '-'"
   }
   validation {
-    condition = alltrue([for o in var.data_disks : (
+    condition     = alltrue([for o in var.data_disks : (
       (o.source_resource_id != null && contains(["Copy", "Restore"], o.create_option) || (o.create_option == "Empty" && o.source_resource_id == null))
     )])
     error_message = "When a data disk source resource ID is specified then create option must be either 'Copy' or 'Restore'."
+  }
+  validation {
+    condition = alltrue([
+      for o in var.data_disks : (
+        (
+          o.disk_iops_read_write == null &&
+          o.disk_mbps_read_write == null &&
+          o.disk_iops_read_only == null &&
+          o.disk_mbps_read_only == null
+        ) || (
+          contains(["UltraSSD_LRS", "PremiumV2_LRS"], o.storage_account_type)
+        )
+      )
+    ])
+    error_message = "disk_iops_read_write, disk_mbps_read_write, disk_iops_read_only and disk_mbps_read_only can only be set for UltraSSD_LRS or PremiumV2_LRS storage account types."
+  }
+  validation {
+    condition = alltrue([
+      for o in var.data_disks : (
+        (
+          o.disk_iops_read_only == null && o.disk_mbps_read_only == null
+        ) || (
+          contains(["UltraSSD_LRS", "PremiumV2_LRS"], o.storage_account_type) &&
+          o.max_shares != null &&
+          o.max_shares > 1 &&
+          o.max_shares <= 10
+        )
+      )
+    ])
+    error_message = "disk_iops_read_only and disk_mbps_read_only can only be set for UltraSSD_LRS or PremiumV2_LRS disks with shared disk enabled (max_shares between 2 and 10)."
+  }
+  validation {
+    condition = alltrue([
+      for o in var.data_disks :
+      o.storage_account_type == "UltraSSD_LRS" ? var.virtual_machine_config.availability_set_id == null : true
+    ])
+    error_message = "UltraSSD_LRS is not supported when 'availability_set_id' is set."
+  }
+  validation {
+    condition = alltrue([
+      for o in var.data_disks : (
+        (o.disk_iops_read_write == null ? true : (o.disk_iops_read_write >= 3000 && o.disk_iops_read_write <= 64000)) &&
+        (o.disk_iops_read_only  == null ? true : (o.disk_iops_read_only  >= 3000 && o.disk_iops_read_only  <= 64000))
+      )
+    ])
+    error_message = "disk_iops_read_write and disk_iops_read_only must be between 3000 and 64000 if set."
+  }
+  validation {
+    condition = alltrue([
+      for o in var.data_disks : (
+        (o.disk_mbps_read_write == null ? true : (o.disk_mbps_read_write >= 125 && o.disk_mbps_read_write <= 750)) &&
+        (o.disk_mbps_read_only  == null ? true : (o.disk_mbps_read_only  >= 125 && o.disk_mbps_read_only  <= 750))
+      )
+    ])
+    error_message = "disk_mbps_read_write and disk_mbps_read_only must be between 125 and 750 if set."
+  }
+  validation {
+    condition = alltrue([
+      for v in var.data_disks :
+      (
+        v.storage_account_type != "UltraSSD_LRS" ? true :
+        try(var.virtual_machine_config.additional_capabilities.ultra_ssd_enabled, false)
+      )
+    ])
+    error_message = "If UltraSSD_LRS is used in data_disks, ultra_ssd_enabled must be set to true in additional_capabilities."
   }
 
   default     = {}
   nullable    = false
   description = <<-DOC
   ```
-   <logical name of the data disk> = {
+   `<logical name of the data disk>` = {
     lun: Number of the lun.
     disk_size_gb: The size of the data disk.
     storage_account_type: Optionally change the storage_account_type. Defaults to StandardSSD_LRS.
@@ -237,7 +355,12 @@ variable "data_disks" {
       the recovery point to restore when create_option is Restore. Changing this forces a new resource to be created.
     write_accelerator_enabled: Optionally activate write accelaration for the data disk. Can only
       be activated on Premium disks and caching deactivated. Defaults to false.
-    on_demand_bursting_enabled: Optionally activate disk bursting. Only for Premium disk. Default false.
+    on_demand_bursting_enabled: Optionally activate disk bursting. Only for Premium disk with size to 512 Gb up. Default false.
+    disk_iops_read_write: (Optional) The maximum number of IOPS allowed for the disk in read/write operations.
+    disk_mbps_read_write: (Optional) The maximum number of MBps allowed for the disk in read/write operations.
+    disk_iops_read_only: (Optional) The maximum number of IOPS allowed for the disk in read-only operations.
+    disk_mbps_read_only: (Optional) The maximum number of MBps allowed for the disk in read-only operations.
+    max_shares: (Optional) The maximum number of VMs that can share this disk. Only for UltraSSD_LRS and PremiumV2_LRS disks.
    }
   ```
   DOC